+++ pkg/commands/oscommands/cmd_obj_runner_default.go
@@ -11,7 +11,7 @@ import (
 
 // we define this separately for windows and non-windows given that windows does
 // not have great PTY support and we need a PTY to handle a credential request
-func (self *cmdObjRunner) getCmdHandler(cmd *exec.Cmd) (*cmdHandler, error) {
+func (self *cmdObjRunner) getCmdHandlerPty(cmd *exec.Cmd) (*cmdHandler, error) {
 	ptmx, err := pty.Start(cmd)
 	if err != nil {
 		return nil, err
